refactor: less code, more reusable and maintenable
- less duplication, less files, less code
- instead of defining endpoints specific to generic and acrobatics, now
  we have block made from generic ocp. We can inherit from these block
  to add behavior (endpoints) or disable some.
  by using classes to define the endpoints.
- removed files (not needed anymore): acrobatics_objectives/constraints,
  {ocp}_config files, read and update are part of new {ocp}_data.py file
  models.py that are not used
- less duplication in requests/responses
- updated tests
- computing acrobatics base data instead of hard-coding it
- minor docs changed

+++ api/bioptim_gui_api/generic_ocp/endpoints/generic_ocp_responses.py
@@ -90,4 +90,4 @@ class ObjectiveFcnResponse(ObjectiveFcnRequest):
 class ArgumentResponse(BaseModel):
     key: str
     type: str
-    value: Union[int, float, str, list, None, Axis]
+    value: Union[float, int, str, list, None, Axis]
